refactor: move Gradle discovery logic from workspace.js to java_gradle.js
Address review feedback: provider-specific discovery functions don't belong
in workspace.js. Extract traverseForWrapper to tools.js as shared utility,
update base_java.js to delegate to it, and move all Gradle workspace discovery
functions (discoverGradleSubprojects, resolveGradleBinary,
parseGradleInitScriptOutput, GRADLE_INIT_SCRIPT) to their provider file.
